+++ src/components/settings/index.ts
@@ -1,5 +1,6 @@
 export { default as GeneralTab } from './GeneralTab';
 export { default as PlaybackTab } from './PlaybackTab';
+export { default as EpgTab } from './EpgTab';
 export { default as ParentalTab } from './ParentalTab';
 export { default as AboutTab } from './AboutTab';
 export * from './constants';